PHPSDK-5

+++ src/Api/TransactionRequests.php
@@ -1,13 +1,14 @@
 <?php
 namespace EnjinCoin\Api;
 
+use EnjinCoin\Util\Numbers;
+use Zend;
 use EnjinCoin\Auth;
 use EnjinCoin\Ethereum;
 use EnjinCoin\EventTypes;
-use PHPUnit\Runner\Exception;
-use Zend;
 use EnjinCoin\ApiBase;
 use EnjinCoin\Util\Db;
+use EnjinCoin\Ethereum as Eth;
 
 class TransactionRequests extends ApiBase {
 	const TYPE_BUY = 'buy';
@@ -47,7 +48,7 @@ class TransactionRequests extends ApiBase {
 	 */
 	public function create(array $identity, array $recipient, string $type, string $icon = null, string $title = null, int $token_id = 0, string $value = '0') {
 		// Validate Txr Type
-		if(!in_array($type, self::$txr_types)) throw new \Exception('Invalid Transaction Request Type');
+		if (!in_array($type, self::$txr_types)) throw new \Exception('Invalid Transaction Request Type');
 
 		// Validate Identity
 		$identities = new Identities();
@@ -66,11 +67,10 @@ class TransactionRequests extends ApiBase {
 		}
 		if (!empty($recip['ethereum_address']) && Ethereum::validateAddress($recip['ethereum_address'])) {
 			$recipient_db['recipient_address'] = $recip['ethereum_address'];
-		}
-		else throw new Exception('Identity does not exist');
+		} else throw new Exception('Identity does not exist');
 
 		// Validate Value
-		if(!Ethereum::validateValue($value)) throw new Exception('Invalid Transaction Request Value');
+		if (!Ethereum::validateValue($value)) throw new Exception('Invalid Transaction Request Value');
 
 		$insert = $this->db->insert('transaction_requests');
 		$timestamp = time();
@@ -82,7 +82,7 @@ class TransactionRequests extends ApiBase {
 			'recipient_id' => $recipient_db['recipient_id'],
 			'recipient_address' => $recipient_db['recipient_address'],
 			'icon' => $icon,
-			'title' => $icon,
+			'title' => $title,
 			'token_id' => $token_id,
 			'value' => $value,
 		], $insert::VALUES_MERGE);
@@ -90,11 +90,15 @@ class TransactionRequests extends ApiBase {
 		$results = Db::query($insert);
 		$txr_id = $results->getGeneratedValue();
 
-		// Create event and notification @todo
+		// Create event
 		(new Events)->create(Auth::appId(), EventTypes::TXR_PENDING, [
 			'txr_id' => $txr_id,
 			'identity' => $identity,
 			'recipient' => $recipient,
+			'type' => $type,
+			'icon' => $icon,
+			'title' => $title,
+			'value' => $value,
 		]);
 
 		return true;
@@ -102,16 +106,27 @@ class TransactionRequests extends ApiBase {
 
 	/**
 	 * Cancel a Transaction Request
-	 * @param $txr_id
+	 * @param int $txr_id
 	 * @return bool
 	 */
-	public function cancel($txr_id) {
+	public function cancel(int $txr_id) {
 		$txr = $this->get($txr_id);
 
-		// Check permissions for cancellation type @todo
+		// Check permissions for cancellation type
+		$event_type = null;
+		if (Auth::role() == Auth::ROLE_SERVER)
+			$event_type = EventTypes::TXR_CANCELED_PLATFORM;
+		else if (Auth::role() == Auth::ROLE_CLIENT)
+			$event_type = EventTypes::TXR_CANCELED_USER;
 
-		$event_type = EventTypes::TXR_CANCELED_USER;
-		$event_type = EventTypes::TXR_CANCELED_PLATFORM;
+		$identity = Auth::identity();
+		$identity_id = $identity['identity_id'];
+
+		if (empty($event_type)
+			|| ($event_type == EventTypes::TXR_CANCELED_USER && empty($identity_id))
+			|| ($event_type == EventTypes::TXR_CANCELED_USER && $txr['identity_id'] != $identity_id))
+			throw new Exception('You do not have permission to cancel this transaction request');
+
 
 		$delete = $this->db->delete('transaction_requests');
 		$delete->where(['txr_id' => $txr_id]);
@@ -120,5 +135,27 @@ class TransactionRequests extends ApiBase {
 		(new Events)->create(Auth::appId(), $event_type, ['txr_id' => $txr_id]);
 
 		return true;
+	}
+
+	public function broadcast(int $txr_id, array $transaction) {
+		$txr = $this->get($txr_id);
+
+		// Validate address
+		if (!Eth::validateAddress($transaction['from']) || !Eth::validateAddress($transaction['to'])) {
+			throw new Exception('Invalid address');
+		}
+
+		// Validate Transaction Request with the Transaction
+		if($transaction['from'] != $txr['from'])
+			throw new Exception('Invalid transaction from');
+		if($transaction['to'] != $txr['to'])
+			throw new Exception('Invalid transaction to');
+		if(Numbers::decodeHex($transaction['value']) != $txr['value'])
+			throw new Exception('Invalid transaction value');
+
+		$model = new Eth;
+		$response = $model->msg('eth_sendTransaction', array($transaction));
+
+		(new Events)->create(Auth::appId(), EventTypes::TX_BROADCASTED, ['txr_id' => $txr_id]);
 	}
 }
